Move EnvCompare::CLI out to it's own file
- Cleans up the behavior by putting it where one would expect to find it
- Adjusts the theme path to compensate
